fix: adjust uuid usage
Multiple changes were made to improve code quality and functionality. Here's a summary of the
changes:

- Debug log statements were removed from `ArticleList.js` and `ArticleSummary.js` for calculation
enabling and disabling methods. This helps to keep the console clean and focussed on critical
issues.
- Changed a misused property name from `hash` to `uuid` in `CopyErpEntityDialog.js`. This ensures
correct usage of identifiers.
- Multiple formatting changes for function declarations in `ERPEntities.js`, functions now follow
the format `function () {}` for unnamed functions.
- A `#id` field is newly included in `openPanelByUUID` function of `ERPEntities.js`.

These changes are expected to augment the performance of ERP modules, and improve the overall
readability of the code.