fix: simplify address selection in UserData.js
This commit simplifies the logic for selecting the current and default address in the `UserData.js`
file. Specifically, it replaces the for-loop with the `Array.find()` method to find the default
address and the current address by `uuid`. If the current address isn't found, it's set to the
default address.

Related: quiqqer/invoice#163